Fixed Add test for org admin approving an organization

+++ app/controllers/admin/organizations_controller.rb
@@ -1,19 +1,29 @@
 class Admin::OrganizationsController < ApplicationController
+
   def approve
     if current_user.is_org_approver?
-      org=Organization.find(params[:id])
-      org.approved=true
-      org.save
-      ApprovalMailer.deliver_organization_approval(org)
+      org = Organization.find_by_id(params[:id])
+      unless org.nil?
+        org.update_attribute(:approved, true)
+        ApprovalMailer.deliver_organization_approval(org)
+        flash[:notice] = "You have approved the #{org.name} organization."
+      else
+        flash[:error] = "The organization does not exist. Has it been previously denied?"
+      end
       redirect_to dashboard_path
     end
   end
   
   def deny
     if current_user.is_org_approver?
-      org = Organization.find(params[:id])
-      ApprovalMailer.deliver_organization_denial(org)
-      org.destroy
+      org = Organization.find_by_id(params[:id])
+      unless org.nil?
+        ApprovalMailer.deliver_organization_denial(org)
+        flash[:notice] = "You have denied the #{org.name} organization."
+        org.destroy
+      else
+        flash[:error] = "The organization does not exist. Has it been previously denied?"
+      end
       redirect_to dashboard_path
     end
   end
